added Repositories (crud only) added/refactored folder structure for future use changed pom to with new dependencies fixed entity table typo in table names uncommented controllers
