+++ src/main/java/com/prideTeam/AstonWebSchool/controllers/student/StudentRestController.java
@@ -1,4 +1,4 @@
-package com.prideTeam.AstonWebSchool.web.student;
+package com.prideTeam.AstonWebSchool.controllers.student;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
