+++ src/main/java/com/prideTeam/AstonWebSchool/entity/entityAbstracts/AbstractBaseEntity.java
@@ -1,4 +1,4 @@
-package com.prideTeam.AstonWebSchool.entity;
+package com.prideTeam.AstonWebSchool.entity.entityAbstracts;
 
 import com.prideTeam.AstonWebSchool.exception.EntityInTransientStateException;
 import org.hibernate.Hibernate;
