+++ src/main/java/com/prideTeam/AstonWebSchool/repositories/AdminRepository.java
@@ -1,0 +1,9 @@
+package com.prideTeam.AstonWebSchool.repositories;
+
+import com.prideTeam.AstonWebSchool.entity.Admin;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AdminRepository extends JpaRepository<Admin,Integer> {
+}
