+++ src/main/java/com/prideTeam/AstonWebSchool/repositories/LessonRepository.java
@@ -1,0 +1,9 @@
+package com.prideTeam.AstonWebSchool.repositories;
+
+import com.prideTeam.AstonWebSchool.entity.Lesson;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LessonRepository extends JpaRepository<Lesson,Integer> {
+}
