+++ src/main/java/com/prideTeam/AstonWebSchool/repositories/LogRepository.java
@@ -1,0 +1,9 @@
+package com.prideTeam.AstonWebSchool.repositories;
+
+import com.prideTeam.AstonWebSchool.entity.Log;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LogRepository extends JpaRepository<Log,Integer> {
+}
