+++ src/main/java/com/prideTeam/AstonWebSchool/repositories/RoleRepository.java
@@ -1,0 +1,8 @@
+package com.prideTeam.AstonWebSchool.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<RoleRepository,Integer> {
+}
