+++ src/main/java/com/prideTeam/AstonWebSchool/repositories/StudentRepository.java
@@ -1,0 +1,9 @@
+package com.prideTeam.AstonWebSchool.repositories;
+
+import com.prideTeam.AstonWebSchool.entity.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface StudentRepository extends JpaRepository<Student,Integer> {
+}
