+++ src/main/resources/application.properties
@@ -1,1 +1,12 @@
+#Spring Data
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://localhost:5432/aston_web_school
+spring.datasource.username=postgres
+spring.datasource.password=post
 
+#Liquibase
+spring.liquibase.enabled=true
+spring.liquibase.change-log=src/main/resources/db/changelog/db.changelog-master.xml
+spring.liquibase.url=jdbc:postgresql://localhost:5432/aston_web_school
+spring.liquibase.user=postgres
+spring.liquibase.password=post
